+++ History.md
@@ -1,4 +1,11 @@
 
+1.1.0 - January 7, 2015
+-------------------------
+API changes: 
+- adding `metric.timestamps()` method
+- adding `metric.values()` method
+- adding `metric.from()` method
+
 1.0.2 - January 6, 2015
 -------------------------
 - fix to allow falsey answers